Refactor code for improved type safety and clarity
- Changed LineOfDutyDao to ILineOfDutyDao for better abstraction.
- Updated calls to LookupService.GetCancelReasonDescription to use Value property for nullable types.
- Cast enum values to int when passing to methods to ensure type consistency.
- Replaced direct parsing of refID with int.Parse for better error handling.
- Removed unnecessary CLSCompliant attribute from LodPrinciple class.
- Modified LoadPermissions method to parse _userName to int for user ID.
- Enhanced TabItemList to use indexed access instead of foreach for better performance.
- Updated ActiveTaskList to use System.Data.DbType for parameter types.
- Changed isFormal property assignment to convert boolean to string.
- Ensured consistent use of System.Data.DbType across workflow-related classes.
- Adjusted WorkflowStepList to include necessary using directive for data access.
- Updated WorkflowViewList to pass row by reference in ToDataRow method.

+++ ALODWebUtility.CS/Documents/CertificationStampStrategy.cs
@@ -148,7 +148,7 @@ namespace ALODWebUtility.Documents
                 return 0;
             }
 
-            LogManager.LogAction(ModuleType.System, UserAction.AddedDocument, "Added Certification Stamped document: " + metaData.OriginalFileName + ". RefId = " + refId + ". DocGroupId = " + groupId + ". DocId = " + docId + ".");
+            LogManager.LogAction((int)ModuleType.System, UserAction.AddedDocument, "Added Certification Stamped document: " + metaData.OriginalFileName + ". RefId = " + refId + ". DocGroupId = " + groupId + ". DocId = " + docId + ".");
 
             return docId;
         }
